Capnproto Pubsub.
